REFACTOR folder structure of client and product
