+++ src/main/java/com/travelport/projecttwo/controllers/dtos/client/ClientRequestDto.java
@@ -1,22 +1,20 @@
-package com.travelport.projecttwo.controllers.dtos;
+package com.travelport.projecttwo.controllers.dtos.client;
 
-import com.fasterxml.jackson.annotation.JsonInclude;
+import jakarta.validation.constraints.NotBlank;
+import org.hibernate.validator.constraints.Length;
 
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public class ClientResponseDto {
+public class ClientRequestDto {
 
-    private String id;
+    @NotBlank
+    @Length(min = 9, max = 10)
     private String nif;
+
+    @NotBlank
+    @Length(min = 2, max = 150)
     private String name;
+
+    @Length(min = 5, max = 150)
     private String address;
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
 
     public String getNif() {
         return nif;
@@ -41,5 +39,4 @@ public class ClientResponseDto {
     public void setAddress(String address) {
         this.address = address;
     }
-
 }
